backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
click==8.3.2
ecdsa==0.19.2
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
idna==3.11
numpy==2.4.4
passlib==1.7.4
pgvector==0.4.2
psycopg2-binary==2.9.11
pyasn1==0.6.3
pydantic==2.13.0
pydantic_core==2.46.0
alembic==1.16.5
python-jose==3.5.0
python-multipart==0.0.26
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
python-dotenv==1.1.1
